Updated - Removed editable actions from the ActionBar when viewing is read-only
